requirements/base.txt

python-dotenv==1.1.1
# whisper==1.1.10
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
sse-starlette==3.0.2
pytz==2025.2


# postgres requirements
sqlalchemy==2.0.43
python-dotenv==1.1.1
alembic==1.16.4
aiosqlite==0.21.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0

httpx==0.28.1
loguru==0.7.3

qrcode==8.2
Pillow==12.0.0

greenlet==3.2.4
bcrypt==4.0.1

rapidfuzz==2.13.7